% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_dataset_path.R
\name{dms_dataset_path}
\alias{dms_dataset_path}
\title{Console command to load dataset files from DMS to DIA-NN}
\usage{
dms_dataset_path(data_package_num, file_type = ".raw", diann_console = TRUE)
}
\arguments{
\item{data_package_num}{(integer) data package ID number.}

\item{file_type}{LC-MS file extension, default is ".raw"}

\item{diann_console}{Option to modify file paths as console command for DIA-NN}
}
\value{
one column data frame.
}
\description{
For saving output for pasting into DIA-NN console, use `readr::write_tsv()` 
with `col_names = FALSE` and `escape = "none"` for a `.tsv` that can be
pasted directly into DIA-NN

Not useful outside of PNNL unless connected through VPN. Works on windows out of the box.
}
\details{
Additional details...
}
\examples{
diann_raw_paths <- dms_dataset_path(5293 , file_type = ".raw")

}
