% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_htrms.R
\name{get_raw_files_from_data_package}
\alias{get_raw_files_from_data_package}
\title{Get full paths to .raw files from a DMS data package}
\usage{
get_raw_files_from_data_package(data_package_num)
}
\arguments{
\item{data_package_num}{(integer) DMS data package ID number}
}
\value{
A character vector containing full paths to all .raw files in the data package
}
\description{
Retrieves the complete file paths to all .raw files associated with a 
  specified DMS data package. This function uses get_datasets_by_data_package() to 
  fetch dataset information and then locates the .raw files in each dataset folder.
}
\details{
This function searches each dataset folder in the specified data package for files
with the .raw extension. It's primarily used as a helper function for 
convert_data_package_to_htrms(), but can also be used independently to get file paths.
}
\examples{
\dontrun{
# Get all .raw file paths from data package 6447
raw_files <- get_raw_files_from_data_package(6447)
}

}
