% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqreport_to_msnset.R
\name{make_esprs}
\alias{make_esprs}
\title{Helper function, makes matrix of expression values}
\usage{
make_esprs(df, id_header, quantity_header, sample_header = "File.Name")
}
\arguments{
\item{df}{data.frame of DIA-NN pq_report}

\item{id_header}{Precursor ID column header}

\item{quantity_header}{Precursors Quantity Column}

\item{sample_header}{Run/Sample header column}
}
\value{
Matrix of quantity_header with col_names = sample_header; row_names = precursor_id,
}
\description{
Helper function, makes matrix of expression values
}
