% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellplate_utils.R
\name{num_to_wellplate}
\alias{num_to_wellplate}
\title{Convert numeric value to wellplate location}
\usage{
num_to_wellplate(num)
}
\arguments{
\item{num}{An integer vector containing numeric values to be converted to wellplate locations.}
}
\value{
A character vector of wellplate locations corresponding to the numeric values.
}
\description{
This function converts numeric values back to wellplate locations in the format "A1", "B2", etc.
where 1 = A1, 2 = A2, 13 = B1, etc.
}
\examples{
num_to_wellplate(c(1, 2, 13, 14, 96))
# [1] "A1" "A2" "B1" "B2" "H12"

}
