% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellplate_utils.R
\name{num_to_wellplate_by_column}
\alias{num_to_wellplate_by_column}
\title{Convert numeric value to wellplate location (by column)}
\usage{
num_to_wellplate_by_column(num)
}
\arguments{
\item{num}{An integer vector containing numeric values to be converted to wellplate locations.}
}
\value{
A character vector of wellplate locations corresponding to the numeric values.
}
\description{
This function converts numeric values to wellplate locations reading down each column.
Format: 1 = A1, 2 = B1, 8 = H1, 9 = A2, etc. Assumes 8 rows (A-H).
}
\examples{
num_to_wellplate_by_column(c(1, 2, 9, 96))
# [1] "A1" "B1" "A2" "H12"

}
