% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellplate_utils.R
\name{num_to_wellplate_by_row}
\alias{num_to_wellplate_by_row}
\title{Convert numeric value to wellplate location (by row)}
\usage{
num_to_wellplate_by_row(num)
}
\arguments{
\item{num}{An integer vector containing numeric values to be converted to wellplate locations.}
}
\value{
A character vector of wellplate locations corresponding to the numeric values.
}
\description{
This function converts the rowwise numeric values back to wellplate locations in the format "A1", "B2", etc.
where 1 = A1, 2 = A2, 13 = B1, etc. Basically Evosep number back to well.
}
\examples{
num_to_wellplate_by_row(c(1, 2, 13, 96))
# [1] "A1" "A2" "B1" "H12"

}
