% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_to_msnset.R
\name{parquet_to_msnset}
\alias{parquet_to_msnset}
\title{Converts Parquet table to MSnSet}
\usage{
parquet_to_msnset(pqt)
}
\arguments{
\item{pqt}{output from `read_diann_parquet()`}
}
\value{
MSnSet object
}
\description{
Converts Parquet table to MSnSet
}
\details{
At this point it return only precurso (pq) type MSnSet. The only additional
piece of data is iRT - normalized retention time of a peptide/precursor.
}
\examples{
\dontrun{
fpath <- system.file("extdata", "report.parquet", package = "PNNL.DIA.utils")
pqt <- read_diann_parquet(fpath)
m <- parquet_to_msnset(pqt) }

}
