% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqreport_to_msnset.R
\name{pqreport_to_msnset}
\alias{pqreport_to_msnset}
\title{Converts DIA-NN precursor report to msnset object}
\usage{
pqreport_to_msnset(
  x,
  id_header = "Precursor.Id",
  quantity_header = "Precursor.Normalised",
  proteotypic_only = F,
  q = 0.01,
  protein_q = 1,
  pg_q = 1,
  gg_q = 1,
  strip_fData = FALSE
)
}
\arguments{
\item{x}{precursor report read into R}

\item{id_header}{protein/peptide/precursor Id column name}

\item{quantity_header}{quantity column name}

\item{proteotypic_only}{only proteotypic peptides and the respective proteins should be considered}

\item{q}{q-value threshold}

\item{protein_q}{uniquely identified protein q-value threshold}

\item{pg_q}{protein group q-value threshold}

\item{gg_q}{gene group q-value threshold}

\item{strip_fData}{TRUE will remove all fData except for id_header, FALSE will attempt to keep protein meta info from DIA-NN.}
}
\value{
MSnSet object
}
\description{
Mirrors functionality of [`diann::diann_matrix()`](https://github.com/vdemichev/diann-rpackage), but outputs an MSnSet object.

If you have multiple DIA-NN precursor reports that you've combined into one 
i.e. `cbind (pq_report1, pq_report2)`) and you want to convert that into one `msnset` object, 
set `strip_fData = TRUE` to remove protein and gene group info columns from the fData slot to avoid errors
when going from the *long* precursor report to the *wide* msnset. Errors emanate from disparities in
protein group assignment for a given precursor in the report. For ex. *precA* in *run 1* could be assigned
to *protein group X*, whereas the same *precA* in *run 2* could be assigned to *protein groups X; Y*. 
When comparing these types of samples, bet set `strip_fData = FALSE` and compare at the precursor level only.
}
\examples{

x_path <- system.file("extdata",
  "diann_report.tsv",
   package = "pnnl.diann.utils")

x <- read_diann_tsv(x_path)

pqreport_to_msnset(x)
}
