% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr_dms_utils.R
\name{preview_trailer_fields}
\alias{preview_trailer_fields}
\title{Preview available trailer fields from a raw file}
\usage{
preview_trailer_fields(raw_path)
}
\arguments{
\item{raw_path}{Path to a raw file (either folder path or full .raw path)}
}
\value{
Character vector of available trailer field names
}
\description{
Preview available trailer fields from a raw file
}
\examples{
\dontrun{
preview_trailer_fields("\\\\\\\\proto-6\\\\AstralZ02\\\\2025_3\\\\dataset_name")
}
}
