% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pq_report.R
\name{read_pq_report}
\alias{read_pq_report}
\title{Reads DIA-NN Precursor report}
\usage{
read_pq_report(path)
}
\arguments{
\item{path}{Filepath to DIA-NN `report.tsv`}
}
\value{
A data frame
}
\description{
Reads DIA-NN Precursor report
}
\examples{
x <- system.file("extdata", "diann_report.tsv", package = "pnnl.diann.utils")
read_pq_report(x)
}
