% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr_dms_utils.R
\name{resolve_raw_path}
\alias{resolve_raw_path}
\title{Resolve a path to a full .raw file path}
\usage{
resolve_raw_path(path)
}
\arguments{
\item{path}{Folder path or full .raw path}
}
\value{
Full path to .raw file
}
\description{
Resolve a path to a full .raw file path
}
\keyword{internal}
