% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diann_tsv.R
\name{sample_basename}
\alias{sample_basename}
\title{Get Basename from Sample LC-MS file path}
\usage{
sample_basename(x, drop_ext = c("no", "last", "all"))
}
\arguments{
\item{x}{Absolute path to LC-MS file}

\item{drop_ext}{Options to drop file extensions: "no", "last", "all", see Details}
}
\value{
basename of LC-MS file path, drops file extension
}
\description{
Get Basename from Sample LC-MS file path
}
\details{
`drop_ext` specifies whether to drop file extensions. Default option is `"no"`, which keeps file exteions; 
`"last"` removes only the last extension, and `"all"` removes  everything after the first period (`.`); this isn't a problem 
if using data from DMS as periods are not allowed in dataset names 
(see [DMS Dataset Naming Conventions](https://prismwiki.pnl.gov/wiki/Dataset#DMS_Dataset_Naming_Conventions));
}
\examples{
x <- "PATH/TO/MSFile.raw.dia"
sample_basename(x, drop_ext = "all")
sample_basename(x, drop_ext = "last")
sample_basename(x, drop_ext = "no")
}
