% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellplate_utils.R
\name{wellplate_to_num_by_column}
\alias{wellplate_to_num_by_column}
\title{Convert wellplate location to numeric value (by column)}
\usage{
wellplate_to_num_by_column(well)
}
\arguments{
\item{well}{A character vector of wellplate locations (e.g., "A1", "B1", "A2").}
}
\value{
An integer vector containing the numeric values corresponding to the wellplate locations.
}
\description{
This function converts wellplate locations to numeric values reading down each column.
Format: A1 = 1, B1 = 2, H1 = 8, A2 = 9, etc. Assumes 8 rows (A-H).
}
\examples{
wellplate_to_num_by_column(c("A1", "B1", "A2", "H12"))
# [1]  1  2  9 96

}
