% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_htrms.R
\name{write_tsv_to_log}
\alias{write_tsv_to_log}
\title{Write dataframe to log file in TSV format}
\usage{
write_tsv_to_log(df, file)
}
\arguments{
\item{df}{(data.frame) The dataframe to write}

\item{file}{(character) Path to the log file}
}
\value{
No return value, called for side effect of writing to log file
}
\description{
Helper function to write a dataframe to a log file in tab-separated format
}
\keyword{internal}
