% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnset_maxLFQ.R
\name{msnset_maxLFQ}
\alias{msnset_maxLFQ}
\title{MaxLFQ quantites of precursors in msnset}
\usage{
msnset_maxLFQ(m)
}
\arguments{
\item{m}{DIA-NN precursors report as MSnSet from `pqreport_to_msnset()`}
}
\value{
list of maxLFQ intensities
}
\description{
MaxLFQ quantites of precursors in msnset
}
