% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmatrix_to_msnset.R
\name{pgmatrix_to_msnset}
\alias{pgmatrix_to_msnset}
\title{Converts DIA-NN matrix report to msnset object}
\usage{
pgmatrix_to_msnset(path, fdata_rownames = "Protein.Group")
}
\arguments{
\item{path}{Filepath to DIA-NN protein or gene group output file}

\item{fdata_rownames}{Which matrix column to be used in fData slot rowname}
}
\value{
MSnSet object
}
\description{
Converts DIA-NN matrix report to msnset object
}
\examples{

x <- system.file("extdata",
  "QC_Mam_19_01_b_DIA_report.pg_matrix.tsv", 
   package = "pnnl.diann.utils")
   
pgmatrix_to_msnset(x)
}
