% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmatrix_to_msnset.R
\name{pgmatrix_to_msnset}
\alias{pgmatrix_to_msnset}
\title{Converts DIA-NN matrix report to msnset object}
\usage{
pgmatrix_to_msnset(x, id_header = "Protein.Group")
}
\arguments{
\item{x}{protein groups matrix read  into R}

\item{id_header}{Which matrix column to be used in fData slot rowname}
}
\value{
MSnSet object
}
\description{
Converts DIA-NN matrix report to msnset object
}
\examples{

pg_path <- system.file("extdata",
  "QC_Mam_19_01_b_DIA_report.pg_matrix.tsv", 
   package = "PNNL.DIA.utils")
   
pg <- read_diann_tsv(pg_path) 
   
pgmatrix_to_msnset(pg)
}
