% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqreport_to_msnset.R
\name{pqreport_to_msnset}
\alias{pqreport_to_msnset}
\title{Converts DIA-NN precursor report to msnset object}
\usage{
pqreport_to_msnset(
  x,
  id_header = "Precursor.Id",
  quantity_header = "Precursor.Normalised",
  proteotypic_only = F,
  q = 0.01,
  protein_q = 1,
  pg_q = 1,
  gg_q = 1
)
}
\arguments{
\item{x}{precursor report read in with read into R}

\item{id_header}{protein/peptide/precursor Id column name}

\item{quantity_header}{quantity column name}

\item{proteotypic_only}{only proteotypic peptides and the respective proteins should be considered}

\item{q}{q-value threshold}

\item{protein_q}{uniquely identified protein q-value threshold}

\item{pg_q}{protein group q-value threshold}

\item{gg_q}{gene group q-value threshold}
}
\value{
MSnSet object
}
\description{
Mirrors functionality of [`diann::diann_matrix()`](https://github.com/vdemichev/diann-rpackage), but outputs an MSnSet object.
}
\examples{

x_path <- system.file("extdata",
  "QC_Mam_19_01_b_DIA_report.pg_matrix.tsv", 
   package = "pnnl.diann.utils")
   
x <- read_diann_tsv(x) 
   
pgmatrix_to_msnset(x)
}
