% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diann_tsv.R
\name{read_diann_tsv}
\alias{read_diann_tsv}
\title{Reads DIA-NN `.tsv` output files}
\usage{
read_diann_tsv(path)
}
\arguments{
\item{path}{Filepath to DIA-NN `.tsv` output file}
}
\value{
A data frame
}
\description{
Reads DIA-NN `.tsv` output files
}
\examples{
x <- system.file("extdata", "diann_report.tsv", package = "pnnl.diann.utils")
read_diann_tsv(x)
}
